+++ scripts/create_registry_schema.py
@@ -325,7 +325,7 @@ keywords = load_preset_keywords()
 # Loop over each schema
 for schema in schema_list:
     # Connect to database to find out what the backend is
-    db_connection = DbConnection(args.config, schema=schema)
+    db_connection = DbConnection(args.config, schema=schema, skip_provenance_reflect=True)
     print(f"Database dialect is '{db_connection.dialect}'")
 
     if db_connection.dialect == "sqlite":